remove some casts which are legacies of originally using float add some casts add comments and where possible explanations about R changes
git-svn-id: https://svn.r-project.org/R/trunk@69158 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/g_cntrlify.c
@@ -104,8 +104,8 @@ attribute_hidden unsigned short *_controlify(pGEDevDesc dd, const unsigned char 
     /* Of the following two words, `fontword' is updated whenever an escape
        sequence like \f0, \f1, \f2 etc. is seen, since raw_fontnum is itself
        updated.  But `symbol_fontword' is fixed */
-    fontword = ((unsigned short)raw_fontnum) << FONT_SHIFT;
-    symbol_fontword = ((unsigned short)raw_symbol_fontnum) << FONT_SHIFT;
+    fontword = (unsigned short)(raw_fontnum << FONT_SHIFT);
+    symbol_fontword = (unsigned short)(raw_symbol_fontnum << FONT_SHIFT);
 
     while (*src != (unsigned char)'\0')
     {
@@ -140,7 +140,7 @@ attribute_hidden unsigned short *_controlify(pGEDevDesc dd, const unsigned char 
                     }
                     if (matched)
                     {
-                        dest[j++] = RAW_ORIENTAL_HERSHEY_GLYPH | (kanji->nelson);
+                        dest[j++] = (unsigned short)(RAW_ORIENTAL_HERSHEY_GLYPH | (kanji->nelson));
                         src += 2;
                         continue; /* back to top of while loop */
                     }
@@ -181,7 +181,7 @@ attribute_hidden unsigned short *_controlify(pGEDevDesc dd, const unsigned char 
                             dest[j++] = RAW_HERSHEY_GLYPH | charnum;
                         else
                             /* a character in one of the fonts in g_fontdb.c */
-                            dest[j++] = (((unsigned short)fontnum) << FONT_SHIFT) | charnum;
+                            dest[j++] = (unsigned short)((fontnum << FONT_SHIFT) | charnum);
                         src += 2;
                         continue; /* back to top of while loop */
                     }
@@ -328,7 +328,7 @@ attribute_hidden unsigned short *_controlify(pGEDevDesc dd, const unsigned char 
                 glyphindex = (src[3] - '0') + 10 * (src[2] - '0') + 100 * (src[1] - '0') + 1000 * (src[0] - '0');
                 if (glyphindex < NUM_OCCIDENTAL_HERSHEY_GLYPHS)
                 {
-                    dest[j++] = RAW_HERSHEY_GLYPH | glyphindex;
+                    dest[j++] = (unsigned short)(RAW_HERSHEY_GLYPH | glyphindex);
                     src += 4;
                     continue; /* back to top of while loop */
                 }
@@ -346,7 +346,7 @@ attribute_hidden unsigned short *_controlify(pGEDevDesc dd, const unsigned char 
                 glyphindex = (src[3] - '0') + 10 * (src[2] - '0') + 100 * (src[1] - '0') + 1000 * (src[0] - '0');
                 if (glyphindex < NUM_ORIENTAL_HERSHEY_GLYPHS)
                 {
-                    dest[j++] = RAW_ORIENTAL_HERSHEY_GLYPH | glyphindex;
+                    dest[j++] = (unsigned short)(RAW_ORIENTAL_HERSHEY_GLYPH | glyphindex);
                     src += 4;
                     continue; /* back to top of while loop */
                 }
@@ -402,7 +402,7 @@ attribute_hidden unsigned short *_controlify(pGEDevDesc dd, const unsigned char 
                         }
                         if (matched)
                         {
-                            dest[j++] = RAW_ORIENTAL_HERSHEY_GLYPH | (kanji->nelson);
+                            dest[j++] = (unsigned short)(RAW_ORIENTAL_HERSHEY_GLYPH | (kanji->nelson));
                             src += 4;
                             continue; /* back to top of while loop */
                         }
@@ -443,7 +443,7 @@ attribute_hidden unsigned short *_controlify(pGEDevDesc dd, const unsigned char 
                                 dest[j++] = RAW_HERSHEY_GLYPH | charnum;
                             else
                                 /* a character in one of the fonts in g_fontdb.c */
-                                dest[j++] = (((unsigned short)fontnum) << FONT_SHIFT) | charnum;
+                                dest[j++] = (unsigned short)((fontnum << FONT_SHIFT) | charnum);
                             src += 4;
                             continue; /* back to top of while loop */
                         }
@@ -470,7 +470,7 @@ attribute_hidden unsigned short *_controlify(pGEDevDesc dd, const unsigned char 
                     }
                 if (matched) /* it's a control code */
                 {
-                    dest[j++] = CONTROL_CODE | i;
+                    dest[j++] = (unsigned short)(CONTROL_CODE | i);
                     continue; /* back to top of while loop */
                 }
             }
